# internal/checkouts/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by unknown module path version unknown version DO NOT EDIT.
package ports

import (
	"time"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Checkout defines model for Checkout.
type Checkout struct {
	Notes        string    `json:"notes"`
	OrderUuid    string    `json:"orderUuid"`
	ProposedTime time.Time `json:"proposedTime"`
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// CreateCheckoutJSONBody defines parameters for CreateCheckout.
type CreateCheckoutJSONBody Checkout

// CreateCheckoutJSONRequestBody defines body for CreateCheckout for application/json ContentType.
type CreateCheckoutJSONRequestBody CreateCheckoutJSONBody